+++ game/board.py
@@ -47,7 +47,7 @@ class Board:
             value = value * multiplier_word
         return value
     
-    def show_board(board): #NEW
+    def show_board(board): 
         print('\n  |' + ''.join([f' {str(row_index).rjust(2)} ' for row_index in range(15)]))
         for row_index, row in enumerate(board.grid):
             print(
@@ -56,8 +56,23 @@ class Board:
                 ' '.join([repr(cell) for cell in row])
         )
 
-    def place_word_in_board(word, location, orientation):
-        pass
+    def get_multipliers(self,row,col): #Fixing
+
+        Lx2 = [(4, 1), (12, 1), (1, 4), (8, 4), (15, 4), (3, 7), (7, 7), (9, 7), (13, 7), (4, 10), (12, 10), (0, 12), (7, 12), (14, 12), (3, 15), (11, 15)]
+        Lx3 = [(6, 2), (10, 2), (2, 6), (6, 6), (10, 6), (14, 6), (1, 8), (5, 8), (9, 8), (13, 8), (2, 10), (6, 10), (10, 10), (14, 10), (6, 14), (10, 14)]
+        Wx2 = [(1, 1), (8, 1), (15, 1), (2, 2), (14, 2), (3, 3), (13, 3), (4, 4), (12, 4), (7, 7), (11, 7), (4, 12), (12, 12), (1, 15), (8, 15), (15, 15)]
+        Wx3 = [(0, 0), (7, 0), (14, 0), (0, 7), (14, 7), (0, 14), (7, 14), (14, 14)]
+
+        for row, col in Lx2:
+            self.grid[row][col] = Square(multiplier=2, multiplier_type='letter')
+        for row, col in Lx3:
+            self.grid[row][col] = Square(multiplier=3, multiplier_type='letter')
+        for row, col in Wx2:
+            self.grid[row][col] = Square(multiplier=2, multiplier_type='word')
+        for row, col in Wx3:
+            self.grid[row][col] = Square(multiplier=2, multiplier_type='word')
+
+
 
 
 
@@ -67,5 +82,4 @@ class Board:
     '''def get_multipliers(row,col)'''
     '''def get_multipliers type(row, col)'''
 
-board = Board()
-print(board.show_board())
+