+++ game/tile.py
@@ -4,5 +4,5 @@ class Tile:
         self.letter = letter
         self.value = value
     
-    def __repr__(self): #NEW
+    def __repr__(self): 
         return f"{self.letter}:{self.value}"